Added sendClientFeatures function
- added sendResourceBalance
- added sendStoreBalance
- changed getClientIcons from 16 to 32

+++ src/player.h
@@ -253,7 +253,7 @@ class Player final : public Creature, public Cylinder
 			return storeInbox;
 		}
 
-		uint16_t getClientIcons() const;
+		uint32_t getClientIcons() const;
 
 		const GuildWarVector& getGuildWarVector() const {
 			return guildWarVector;
